minor fix in Transfer

+++ inc/comm/Transfer.h
@@ -46,7 +46,7 @@ public:
   
   Transfer( tRawData data_ );
   Transfer( tRawData header_, tRawData data_ );
-  Transfer( tRawData header_, const uint8_t* pData_, uint16_t dataLength_ );
+  Transfer( tRawData header_, const uint8_t* pData_, size_t dataLength_ );
   
   virtual ~Transfer();
   
@@ -67,7 +67,7 @@ public:
 //  uint8_t* getDataPtr() const { return m_pData.get(); }
 //  void setData( const uint8_t*, uint16_t );
   const tRawData& getData() const { return m_data; }
-  void setData( const uint8_t*, uint16_t );
+  void setData( const uint8_t*, size_t);
   
   size_t size() const noexcept{ return m_data.size(); }
